The object NetworkData now can be executed for parts.
New functions added, save_to_disk and load_from_disk. Now
the object NetworkData itself takes care with the whole process
of saving and loading results.
The Keras model will be saved as a HDF5 file.
Also the function for normalize the acivations of a neuron has
has been moved.