Big optimizations on Orientation and Symmetry Indexs.
SpeedUp of 6x in symmetry indexs and I don't look it, but greater on orientation Index (specially for lower degrees that requires more processing)
Optimization based on take profit of cache (having all calculs from the same image sequentially), and packing all the data for send it to the GPU sequentially and in major batches. Bucles reduced in order to only have one instruction and one asigment in per axis (or per rotation) inside the more internal bucle. All copys maded first an same time and assignments of matrix maded in place.

Now, read_activations.py hacve a Global variable "ACTIVATIONS_BATCH:_SIZE" that tells how much images process in one gpu step. 200 is a good size for images of 224*224*3, if gpu have greater VRAM this number can be greater and the calculs will speed up